requirements.txt
fastapi==0.72.0
uvicorn==0.17.0
cassandra-driver==3.25.0
python-dotenv==0.19.2
email-validator==1.1.3
argon2-cffi==21.3.0
pytest==6.2.5
jinja2==3.0.3
